Modify IdModel Loop map to be aware of `stage_slice_position` (#4502)
For ping-pong warp specialized kernels, the `compute_at_position` must
be to the left of the warp-specialized thread axis because of parallel
type propagation rules. If not, the warp-specialized thread axis will
incorrectly appear in the `AsyncWarp`. For example, see
`TEST_F(PingPongCircularBuffering, ProducerWarpSpecializedError)`.

The `compute_at_position` is defined from producer to consumer. Thus,
sibling TMA load operations cannot be inlined unless they are inlined
with their consumer.

This PR uses the `stage_slice_position` to inline AsyncWarp operations
together. #4460 added the concept of `stage_slice_position`, which
defines the shape of circular buffer stage and how the `mbarrier`
synchronization is inserted in `AsyncWarp` and `ComputeWarp`.

## Detailed changes:
* In the circular-buffer analysis pass, `createAsyncWarps` groups
warp-specialized async operations into separate warps. At the moment,
**only one warp** is created but this will change with multi-role
specialization for Blackwell.
* In IdModel, `buildAsyncWarpInliningInfo` updates `ordered_sibling_ids`
and `sibling_maps` in `StatefulInliningInfo` to create the sibling
relationship in the LOOP graph.
* Created `TEST_P(SiblingPingPongCircularBuffering, TwoTmaLoads)` to
test inlining two sibling TMA loads based on `stage_slice_position`.
* The tests only works when IdModel is enabled. #4555 updates the
`ComputeAtMap:: allocateIndexVariables` to remove the `GTEST_SKIP`.

## Example: `tv2 = add(tv0, tv1)`
The fusion's schedule loads `tv0` and `tv1` to shared memory with TMA
and creates a warp specialized pipeline using the `TIDy` thread axis.

Let the iterDomain schedule for the TensorViews be:
* `tv0_smem[outer_persistent, serial(2), bulk(128)] = tma_load(tv0)`
* `tv1_smem[outer_persistent, serial(2), bulk(128)] = tma_load(tv1)`
* `tv2[outer_persistent, TIDy(2), TIDx(128)] = add(tv0_smem, tv1_smem)`

Let the `compute_at_position` be 1 because the tma loads and add
operation share the outer-most for-loop. Let the `stage_slice_position`
be 2, so the kernel loads 128 elements of tv0 and tv1 asynchronously and
compute two warp
groups in parallel.

The CTA shape is `(TIDx = 128, TIDy = 3, TIDz = 1)`.

Here is the Warp-Specialized kernel structure:
```python
if threadIdx.y == 2: # AsyncWarp
  for idx in range(outer_persistent):  # compute_at_position(1)
    for jdx in range(2):  # stage_slice_position(2)
      mbarrier::wait(empty)
      mbarrier::arriveExpectTx(full)
      tv0_smem = tma_load(tv0)
      tv1_smem = tma_load(tv1)
else:
  for idx in range(outer_persistent): # compute_at_position(1)
    for TIDy(2): # stage_slice_position(2)
      mbarrier::wait(full)
      tv2 = add(tv0_smem, tv1_smem)
      mbarrier::arrive(empty)
```