Modify ComputeAtMap to be aware of stage_slice_position (#4555)
This PR updates `ComputeAtMap::allocateIndexVariables` to allocate the
same index variable for sibling operations in the `AsyncWarp`. It is
stacked on top of #4502.

## Details
1. Gather all async operations using `createAsyncWarps`.
2. For each TensorView, map sibling iterDomains to the left of the
`stage_slice_position` to the same `ValGroup`. See
`getAsyncWarpSiblingIds` and `getSiblingIds`.
3. In `ComputeAtMap::allocateIndexVariables`, determine if the
`concrete_loop_id` is associated with any async operation.
4. If so, find the corresponding loop variable for that `iterDomain`.
Create one if it doesn't exist. Otherwise, reuse the loop variable for
all sibling iterDomains.
5. `async_warp_sibling_id_index_variable` tracks the loop variable for
each `ValGroup`.