Backward contiguity analysis (#4553)
This PR extends the contiguity analysis to support backward traversals
as well. Previously, the contiguity analysis was mostly just adapted
from the legacy indexer and was only able to find contiguous IDs through
forward traversals.

This extension allows for simpler indexing when there's both forward and
backward traversals from allocation/logical domains to loop domains. For
example, consider a case as illustrated below:


![Scratch-39](https://github.com/user-attachments/assets/b7f23870-170a-4895-8674-6238d57990bc)

(See `ContigIndexingTest.Transpose` for the complete Fusion setup)

Here, the colors represent exact mappings. The logical domains of T1 ,
`2` and `3`, are transposed to form `5` and `4` of T2. Both of them are
split, and their inner outputs are further transformed. Both T1 and T2
appear to be scheduled in the same way, but because of the transpose,
when indexing T1 as the consumer of T2, the indexing traversal starts
from the T2 leaf IDs (`22`, `23` and `21` as well as `17` and `15`),
traversing upwards to the matching IDs such as `18` and `16`, then
traversing down to the leaf IDs of T1 (here, T1 allocation domain is
assumed to be its loop domain).

In this case, the ID 10 is contiguous, so instead of traversing further
down to `13`, `14` and `12`, it should be valid to stop at `10`. That is
not what happens with the current TOT as the contiguity analysis only
traverses to the forward direction from the allocation domain to loop
domain, so it knows nothing about `10`.

NOTE: As part of experimentation with Cursor, I extensively used it to
create this PR. Some part of the code is mostly automatically written by
Cursor.