Map CGA dims using 3D grid in persistent Hopper matmul (#4546)
This is a follow-up to #4405 which uses the CGA split to properly
parallelize the resulting CGA dims.

Given something like `cluster_dims = {4, 2, 1}`, this PR creates a 3D
grid of size `{4, 2, 16}` (assuming 132 SMs). This makes it trivial to
launch the proper number of CTAs to guarantee a single wave and it can
simplify indexing for finding the position within the CGA. That is not
currently needed but it would be if we want to do explicit multicast or
broadcasting within a CGA.

TODO: a test exercising 2D CGA with both rasterization orders,
persistent and non-persistent.