CUB-based block-parallel argsort implementation as a device func (#4602)
This PR adds a block-parallel device function of argsort using CUB. 

See
[here](https://docs.google.com/document/d/1vd2YQtOPqvhFQuHu0BNM7X1F8HKoOUkq1f5yNFwTnyo/edit?tab=t.0#heading=h.mkxhhy66h6e)
for the background.

This PR only adds the device function and unit tests. The device
function is not yet used from generated code, which will be done in a
follow-up (#4598).

Tried to follow the same pattern as the existing device routines such as
`blockReduce`, but there are some major limitations:

- CUB requires static CTA shapes. Probably not an immediate concern,
though.
- Unlike `blockReduce`, it is not possible to just sort a certain
threading dimension, say `TIDx`. I think we could support that by
wrapping values to sort with thread indices, but not implemented.
- Haven't checked yet, but unlikely to support diverged threads, i.e.,
all threads would need to call the function in the exact same call
stack.