Experimental argsort codegen support (#4598)
Stacked on #4602 

This PR adds minimal changes to enable an `ArgsortOp` with the device
func added in #4602. See `ArgsortTest.BasicExecution`.

There are still many limitations. The lowering pass is not taken by
default as the scheduler always uses the fallback path, so at this
moment, this is considered experimental only and as such I didn't put
much effort for safe guarding.

One big WAR is the use of CUB. When `ArgsortOp` is used, we
automatically add `-I/usr/local/cuda/inlude` for the CUB header path.
This obviously won't work unless the execution machine has the
compatible CUB headers in that location. For now, however, this should
be good enough for our experimentation.

Another major change is the use of namespaces in generated code. There
are a couple of things to consider:

- argsort.cu requires cub/block/block_radix_sort.cuh, so it's most
natural to have `#include <cub/block/block_radix_sort.cuh>` in its file,
which means that we can no longer automatically put argsort.cu in an
namespace like [we current
do](https://github.com/NVIDIA/Fuser/blob/main/csrc/runtime/compiled_kernel.cpp#L1128).
Instead, argsort.cu itself needs to have a namespace, which can
technically be anonymous.
- `blockArgsort` uses some utility functions defined in index_utils.cu.
Since `blockArgsort` is in a namespace by itself, the util functions of
index_utils.cu cannot be in an anonymous namespace.
- We only want to include argsort.cu in generated code when necessary.

Overall, what I did is to put everything in the `nvf` namespace. The
kernel part looks like below:

```
namespace nvf {

// Codegen generated code
__global__ void nvfuser_none_f0_c0_r0_g0(Tensor<__half, 2, 2> T0, Tensor<int64_t, 2, 2> T3) {
  nvfuser_index_t i0;
  i0 = ((nvfuser_index_t)threadIdx.x) + (8 * ((nvfuser_index_t)blockIdx.x));
  Array<__half, 1, 1> T1;
  T1[0] = 0;
  T1[0]
     = T0[i0];
  Array<int64_t, 1, 1> T2;
  argsort::blockArgsort<8, 1, 1, 0, 1, 1, __half, 1>(*(int64_t(*)[1])(T2.array + 0), *(__half(*)[1])(T1.array + 0), false, DefaultBlockDim());
  T3[i0]
     = T2[0];
}

} // namespace nvf
```