Concretize K of TopKOp (#4629)
`TopKOp` can have a symbolic IterDomain. This PR extends the fusion
concretizer for `TopKOp`.

Pretty much all of the code was written by Cursor. The change of
logical_domain_map.cpp was copied from #4620 just because I already
fixed it in the PR branch, but Cursor was able to identify the issue
with the pairwise mapping and presented a right fix, which I think quite
impressive.