Some adjustments on Ampere matmul (#4675)
- SASS tests are skipped in whole. The disassembled cubin string format
has no compatibility guarantee, so these tests might fail.
- Allow running ampere matmuls on all devices: Ampere's `mma`
instruction is available on all devices, not just Ampere. `mma` on
Hopper/Blackwell is not as fast as `wgmma` or `tcgen05.mma`, but is
functionally available and can deliver Ampere-level performance. Ampere
tests are still skipped on Hooper & Blackwell, to save CI time.