Fix condition for smem limit in kernel executor (#4644)
Previously we checked that the required static+dynamic smem of our
kernels are < the device limit. This causes problems on some unlucky
matmul problem sizes were we'd like to use the entire range of shared
memory. This PR changes that condition to <=.

I also added a test that tries to exercise this and noticed that the
kernel required 16 bytes of static smem surprisingly.