ScanOp with Aten fallback (#4699)
Contest: prefix-sum is used to reorder the inputs of grouped GEMM in
MoE.

Stole some of the code from #4261 to create a bare minimum support of
PyTorch's `cumsum`. Omitted some of the additional parameters like
exclusive scan and discount factor as it isn't necessary (yet).

In this PR, only the `FusionExecutorCache` path is considered, which
always uses the ExprEval scheduler, so lowering is not considered. A
follow-up PR will address lowering by following the black-box approach.