Allow the quick promotion analysis even with broadcast and non-broadcast mixed loop groups (#4822)
This PR fixes the following error with the added repro:

```
Error from segmentation group 1:  INTERNAL ASSERT FAILED at "/opt/pytorch/nvfuser/csrc/val_graph_visitor.h":123, please report a bug with repro script to NVFuser at https://github.com/NVIDIA/Fuser/issues. The graph has an infinite loop. The following Vals should be visited but are never ready:
```

This happens in the loop promotion analysis, which assumes the graph is
acyclic. Cycles are unavoidable with patterns like RoPE, at least with
the current definition of the Exact graph, for which we are using the
quick
[shortcut](https://github.com/NVIDIA/Fuser/blob/main/csrc/id_model/loop_promotion.cpp#L1255)
whenever possible. The shortcut can only be used with limited cases
where the broadcast forwarding isn't used.

The error happened because the shortcut wasn't used even though there's
no broadcast forwarding. The logic of determining the validity of using
the shortcut was too restrictive. This PR relaxes it so that the repro
case can be handled by the shortcut rather than the full loop promotion
analysis, which failed due to the cycle in the graph.

My intention has been that we always use the shortcut for RoPE-like
patterns with the resize scheduler, but apparently it wasn't the case.
For example, the hf_qwen2 RoPE benchmark failed to use the shortcut but
still didn't fail at the full loop promotion analysis even thought its
exact graph did have a cycle. I suspect it worked fine just because the
analysis does have some
[tolerance](https://github.com/NVIDIA/Fuser/blob/main/csrc/val_graph_visitor.cpp#L51-L86)
of simple cycles for the AlmostExact graph.