A simple implementation of Llama 4's MoE layer (#4830)
This is what an ideal model definition may look like for nvFuser. Unlike
the HF implementation, this implementation uses argsort, scatter and
gather to keep activations small. It's yet to use `torch._grouped_mm`
due to the offsets-must-be-multiples-of-16 constraint from
`torch._grouped_mm`. For that, I plan to change nvFuser's GroupedMmaOp
fallback to run gemms individually. This won't be SOL obviously, but
will allow us to use `torch._grouped_mm` in the model definition and run
it via Thunder/nvFuser.

cc @kshitij12345