Experimental scan codegen support (#4776)
Stacked on top of #4775 

As is done for TopKOp and Argsort, this is not yet piped through
FusionExecutorCache, so no use case should be affected. The lowering of
ScanOp is still quite limited but is at least sufficient for the new
test cases. We'll make it incrementally more robust and flexible.