Move host IR lowering to host_ir/lowering.h (#4956)
In addition to the code motion, this PR minimizes the information
required to pass from FusionKernelRuntime to
lowerSegmentedFusionToHostIr:
1. group_run_order is recomputed inside lowerSegmentedFusionToHostIr. 
2. During parallel compilation, KernelExecutors are added to
`executors_` and then moved to `HostIrContainer`. This way,
HostIrContainer can be created entirely inside lowerSegmentedFusion,
compileKernel doesn't need to be bifurcated, and HostIrContainer doesn't
need a num_groups argument.
3. CompileParams from CompiledKernel are used, so we don't need another
argument to pass in CompileParams.