Allow the inner persistent scheduler to schedule nvfp4 quantization kernels (#4951)
Extend the support @zasdfgbnm had added to nvFuser to support
quantization to nvfp4 in the inner persistent scheduler.
This follows the support added to the pointwise scheduler to vectorize
casts added in the [PR.](https://github.com/NVIDIA/Fuser/pull/4695).

We need to used the work around in the
[PR](https://github.com/NVIDIA/Fuser/pull/4831/) to make sure casts
other than those involved down cast to nvfp4 are not vecotorized.

The segments after this
PR:https://gist.github.com/protonu/674e71480d456d856cb691392b474d2a
Prior to this PR, there would be 3 segments - reduction, pointwise and
expr_eval.