Add a default .pre-commit-config.yaml
Running lintrunner pre commit reduces the time developers spend fixing lint errors in CI.

This was initially suggested by
https://github.com/NVIDIA/Fuser/pull/958#issuecomment-1738326868, but I
had never got a chance to add it.

I don't think this config is highly environment-specific, but let me
know if you prefer not checking in the config. As a reference, Thunder
has .pre-commit-config.yaml checked in as well.