[Refactoring] Decoupling intermediate states from GpuLower (#4980)
Motivation: Some of the analyses of lower_device/analysis can be used
outside the lowering, but they are tightly bound with `GpuLower`, which
is inconvenient when used outside of lowering. For example, I'm
considering using `SyncMap` for the greedy scheduler, but there's no
easy way to use it as `SyncMap` uses some other analysis results and
they are assumed to be accessible through `GpuLower::current()`, which
is not the case if used outside of lowering.

I tried to decouple them from `GpuLower` so that `SyncMap` is no longer
dependent on `GpuLower`. The analysis results that are currently held in
`GpuLower` are moved to a new class called `FusionInfo`, which works as
a container of these analysis results. Much like `GpuLower::current()`,
`FusionInfoGuard::current()` can be used to make dependent results
accessible as long as a valid current `FusionInfo` is in a
`FusionInfoGuard`.

In this PR, I only updated only those necessary for `SyncMap`. If this
looks good, we should incrementally update all of the remaining ones to
follow this convention.

Something I haven't addressed is the dependency management of the
analyses. For example, to build a `SyncMap`, there are six dependent
analyses that need to be done in a particular order
(https://github.com/NVIDIA/Fuser/pull/4980/files#diff-9f148c6abf4059e23d7b0bb5f76736b320a711f920e899836e15cfa2e5ece398R3109-R3117).
This should be automatically taken care, i.e., building a `SyncMap`
automatically triggers dependent analyses if not done yet.