Enable vectorization within reduction schedulers for sharded tensorviews (#5001)
For #4873 and #2563.

PR https://github.com/NVIDIA/Fuser/pull/4873 exposed gaps within
reduction scheduler which were affecting performance for sharded
tensorviews.

When the inner dimension is sharded, the vectorization analysis failed
and produced a vectorization_factor of 1.
This PR:

1. Modifies `getReductionProperties` to record per-GPU sizes instead of
global sizes -> this avoids launching kernels with grid and block size
based on global sizes, which led to idle threads.
2. Modified `getContigMergeOfInnerSize` to be multidevice-aware:
`ContiguousInnerDimensionsMapper` records extents of logical sizes and
traverses through MaxInfoSpanningTree. This previously led to mismatch
between the project dimensions found and the allocation ids of tensors.
3. @liqiangxl identified a bug in `FindAllMappedDims`:
`FindAllMappedDims` was attempting to use allocation ids in
PairwiseLogicalDomainMap and was missing some inner iterdomains as a
result. The fix is to project back to logical for using the pairwise
map. This was the also the reason for #4986. At this time, I modified
this projection to only expect DID loop splits (See code comment).

---------
Co-authored-by: Liqiang Lu <liqiangxl@gmail.com>