Greedy scheduler (#5014)
This is an initial version of the greedy scheduler. This is still
preliminary with a number of limitations (#5030). However, it is almost
sufficient for fusions like `SgLangMoeTest.ComputeExpertOffsets`, except
for reshape, which will be addressed in a follow-up PR.

The scheduler is not yet enabled by default.
`NVFUSER_ENABLE=greedy_scheduler` is required.

Stacked on #5053 
Closes #5031

---------

Co-authored-by: jjsjann123 <jiej@nvidia.com>