Add support for scatter to the greedy scheduler (#5084)
This PR adds support of scatter in the greedy scheduler. I initially
thought only the output tensor would be marked as constrained, but all
of the inputs ended up constrained. Additionally, the consumer tensor of
the output tensor needs to be marked as well. Please see the comments in
greedy.h.

One major limitation is allocation domain. Currently, I only considered
the most simple case, where the logical domain is just used as is. It
would be necessary to make it more flexible, but not addressed in this
PR.

Closes #5033

---------

Co-authored-by: jjsjann123 <jiej@nvidia.com>