Introduce Cutlass codegen utility function (#5087)
This is the first in a series of PRs that will enable the new Cutlass
scheduler. This first PR only introduces the function that converts a
Fusion to a CUTLASS code along with a launcher function. The included
test currently just tests that the code runs but will be modified in the
future to test execution of the kernel.

---------

Co-authored-by: Ryan Spring <rspring@nvidia.com>