Support swizzled output allocation domain when quantizing to nvfp4 (#4952)
The goal of this PR was to demonstrate that our NVFP4 quantization
kernels can support an allocation domain on the output block scaling
factors. To do so we also had to modify the logic in our runtime where
we allocate memory for our outputs.

1. This PR adds a test where the block scales (output) of quantization
has a padded, swizzled allocation domain
2. Modifies inferring output shapes to use `as_strided` instead of
`view` when modifying aten tensors while traversing between allocation
domain and logical domain.