Fix predicate elimination with shared memory tensors (#5107)
The predicate elimination check with shared memory is skipped when
there's no producer tensor, which is wrong. This happens, for example,
with factory ops. Please see the new test.

Here's the generated kernel with the repro using ToT:

```
__global__ void nvfuser_none_f0_c0_r0_g0(Tensor<float, 1, 1> T0, Tensor<float, 1, 1> T1, Tensor<float, 1, 1> T3) {
  alignas(16) extern __shared__ char array[];
  const unsigned smem_offset = 0;
  float* T2 = reinterpret_cast<float*>(array + smem_offset + 0);
  float f0;
  f0 = __to_float(1);
  float f1;
  f1 = __to_float(1);
  T2[((nvfuser_index_t)threadIdx.x)] = 0.000000000e+00f;
  if ((((nvfuser_index_t)threadIdx.x) < 8)) {
    T3[((nvfuser_index_t)threadIdx.x)]
      = T2[((nvfuser_index_t)threadIdx.x)]
      + f0;
  }
  T1[((nvfuser_index_t)threadIdx.x)]
    = T0[((nvfuser_index_t)threadIdx.x)]
    + f1;
}
```

Notice that the predicate of the zero assignment to `T2` is eliminated,
which is wrong.

~TODO: It assumes the allocation of a shared memory tensor is always
based on the loop domain, which may not be true.~ Allocation domain
should not matter for synchronization.

---------

Co-authored-by: Jacob Hinkle <1454944+jacobhinkle@users.noreply.github.com>