Ignore different symbol names in codediff (#5106)
This fixes false positives in codediff such as [this
example](http://nv/e5M/nvfuser_github_ci/codegen_diff_p34324383_j206130253_1756904251218282579_codediff_c490444f_532ddcc8_custom_command_20250903_051343.html).

A diff like this should not be counted as an actual code diff and with
the PR it won't be:
```diff
--- c490444f
+++ 532ddcc8
@@ -16,35 +16,35 @@

 .global .align 1 .u8 _ZN3nvf3std17integral_constantIbLb1EE5valueE;
 .global .align 1 .u8 _ZN3nvf3std14__numeric_typeIvE5valueE;
 .global .align 1 .u8 _ZN3nvf3std9is_same_vINS_15DefaultBlockDimES2_EE = 1;
 .extern .shared .align 16 .b8 _ZN3nvf5arrayE[];
 
-.visible .entry _ZN3nvf12nvfuser_1085ENS_6TensorIfLi2ELi2EEENS0_IfLi1ELi1EEE(
-	.param .align 8 .b8 _ZN3nvf12nvfuser_1085ENS_6TensorIfLi2ELi2EEENS0_IfLi1ELi1EEE_param_0[24],
-	.param .align 8 .b8 _ZN3nvf12nvfuser_1085ENS_6TensorIfLi2ELi2EEENS0_IfLi1ELi1EEE_param_1[16]
+.visible .entry _ZN3nvf12nvfuser_1083ENS_6TensorIfLi2ELi2EEENS0_IfLi1ELi1EEE(
+	.param .align 8 .b8 _ZN3nvf12nvfuser_1083ENS_6TensorIfLi2ELi2EEENS0_IfLi1ELi1EEE_param_0[24],
+	.param .align 8 .b8 _ZN3nvf12nvfuser_1083ENS_6TensorIfLi2ELi2EEENS0_IfLi1ELi1EEE_param_1[16]
 )
 {
 	.reg .pred 	%p<35>;
 	.reg .f32 	%f<79>;
 	.reg .b32 	%r<169>;
 	.reg .b64 	%rd<35>;
```
We do this by replacing words starting with `_` such as
`ZN3nvf12nvfuser_1085ENS_6TensorIfLi2ELi2EEENS0_IfLi1ELi1EEE_param_0`
with `_foo` when diffing. In the codediff report, the original PTX is
still included, not the substituted text.