PR2: fixing project from allocation to logical for non-divisible split
This PR changes the projection to ensure extent for non-divisible split on backward projection maintains the correct extent, instead of including padded sizes from the two split IDs.

e.g. with ID i0 (extent 5), if we do split by 4, we'll get two new ID

    i0 (extent 5)
   /  \
 i0/4  4

Because i0/4 is a ceilDiv, it will ended up having extent 2.
When we project the extent from i0/4 and 4 back, the combined extent will be 8, instead of 5.

When we allocate buffer, we want the buffer to use codomain sizes (buffer size should be 5). But the tensor we produce should have size that's faithful to its logical domain sizes. Otherwise, later consumer of the tensor would assert on mismatch tensor sizes.