Correctly compute K in cutlass executor's nvfp4 scaled GEMM (#5182)
This PR fixes the test in `test_cutlass_scheduler.cpp` to actually
generate random values properly. In doing so, I noticed that the
previous version still used `at::empty` instead which commonly used all
zeros. This masked a bug that is also fixed in this PR where we computed
the K dimension improperly.

---------

Co-authored-by: Ryan Spring <rspring@nvidia.com>