Refactor lowering of ParallelType::Group (#5204)
The current way of passing inputs and outputs to grouped runtime
functions seems to be fragile. See a concrete example at #5206. We can't
just pass, e.g., `T2.array`, unless `T2` is known to be allocated with
the grouped iter domain only. I didn't find any validation, and the
repro of #5206 doesn't cause any assertion but just silently produces
incorrect outputs.

Also, a for-loop is not considered grouped unless there are specific ops
inside. In other words, grouping may be just simply ignored otherwise,
which I don't think is a good interface.

Main changes of this PR:

- Forloop with a grouped iter domain is always grouped. This is fine for
those runtime ops for which the grouping is intended to be used, but
LoadStoreOp also needs to be able to work with grouped tensors as some
tensors may need to be initialized (e.g., reduction output tensors).
- Added kir::GroupedLoadStoreOp for the initialization of grouped
tensors
- Changed codegen.cpp to pass proper offsets to input and output buffers

There's a still problem of allocation ordering, which I'll address in a
next PR. For now, just added a repro (disabled).

Fixes #5206