[Refactoring] Change cacheInputs to return pairs of a cache and the original input (#5271)
This is just minor refactoring with no logic change. For `cacheInputs,`
needed to access both newly created caches and their corresponding
inputs, so updated it to return a vector of tensors much like
`cacheAndForkOutputs`. In addition, since just returning a pair of
tensors can be confusing as you'd need to remember which tensor
corresponds to which, changed both `cacheInputs` and
`cacheAndForkOutputs` to return a vector of pairs of cache tensors and
indices of corresponding fusion inputs (or outputs).