Make sure the work buffer can be safely reused (#5252)
CUB routines do not automatically sync with respect to the shared memory
work buffer at the end. Repeated calls to the same routine may end up
causing a write after read hazard. This PR fixes the issue by inserting
a syncthreads at the end of each CUB-based function. We could omit it in
some cases, for example, if a global function only uses it once, it
should never be a problem. We'll revisit as necessary.