Fix transpose scheduler with dangling broadcast IDs (#5269)
Fixes #4957 

This
[fusion](https://github.com/NVIDIA/Fuser/issues/4957#issuecomment-3189703218)
results in an misaligned read access to the shared memory tensor, `T18`.
The read happens when storing from `T18` to `T8`.

```
T18_s___bfloat[iblockIdx.x169{256}, iUS170{1}, iS363{4}, ithreadIdx.x364{128}, iS362{2}, bS31{1}] ca_pos( 2 ) produce_pos( 5 )
 logical domain : (bS30{1}, bS31{1}, iS32{4096}, iS33{64})
 allocation domain : (bS30{1}, iS32{4096}, iS33{64}, bS31{1})
 contiguity: n t t n
  Split: iS33{64} by factor 32 -> iS165{2}, iS166{32}
  Split: iS32{4096} by factor 32 -> iS163{128}, iS164{32}
  Merge: iS166{32} and iS164{32} -> iS360{1024}
  Merge: iS163{128} and iS165{2} -> iS167{256}
  Split: iS360{1024} by factor 2 -> iS361{512}, iS362{2}
  Merge: bS30{1} and iS167{256} -> iS168{256}
  Split: iS168{256} by factor 1 -> iblockIdx.x169{256}, iUS170{1}
  Split: iS361{512} by factor 128 -> iS363{4}, ithreadIdx.x364{128}
 loop domain : (iblockIdx.x169{256}, iUS170{1}, iS363{4}, ithreadIdx.x364{128}, iS362{2}, bS31{1})
T8_g___bfloat[iblockIdx.x177{256}, iUS178{1}, iS172{32}, iUR238{1}, ithreadIdx.x239{128}, iV237{2}] ca_pos( 3 ) produce_pos( 2 )
 logical domain : (bS65{1}, bS66{1}, iS67{4096}, iS68{64})
 allocation domain : (bS65{1}, iS67{4096}, iS68{64}, bS66{1})
 contiguity: n t t n
  Split: iS68{64} by factor 32 -> iS173{2}, iS174{32}
  Merge: iS174{32} and bS66{1} -> iS235{32}
  Split: iS67{4096} by factor 32 -> iS171{128}, iS172{32}
  Split: iS235{32} by factor 2 -> iS236{16}, iV237{2}
  Merge: iS171{128} and iS173{2} -> iS175{256}
  Split: iS236{16} by factor 128 -> iUR238{1}, ithreadIdx.x239{128}
  Merge: bS65{1} and iS175{256} -> iS176{256}
  Split: iS176{256} by factor 1 -> iblockIdx.x177{256}, iUS178{1}
 loop domain : (iblockIdx.x177{256}, iUS178{1}, iS172{32}, iUR238{1}, ithreadIdx.x239{128}, iV237{2})
```

In typical normalization kernels, where tensors are grouped into two
disjoint sets, both of them follow a similar pattern but with a
differently ordered merge. Here, `T18` looks normal, but `T8` doesn't.
The most obvious strange scheduling is:

```
Split: iS236{16} by factor 128 -> iUR238{1}, ithreadIdx.x239{128}
```

This would create 128 threads but only first 16 threads would be active.

`T8` is a tensor included in the second tensor group, and in this case
the first group of tensors don't have any corresponding ID of `bS66`.
This results in leaving it unscheduled even after the initial splits and
merges with the first group. And in this particular case it happens to
be located at the innermost of the loop domain of `T8`, which then
causes this
[merge](https://github.com/NVIDIA/Fuser/blob/main/csrc/scheduler/transpose.cpp#L1053),
which is supposed to merge the two tile IDs, to merge the broadcast ID
with one of the tile ID, i.e.,

```
Merge: iS174{32} and bS66{1} -> iS235{32}
```

This is because at the time of the merge, the loop domain looks like
`[..., iS172{32}, iS174{32}, bS66{1}]`. Instead, we should do the merge
as `reference2->merge(-3, -2)`.

This PR avoids this problem by making sure such dangling broadcast IDs
are moved to the innermost positions and ignoring them in the rest of
the scheduling. After this change, `T8` is scheduled as:

```
T8_g___bfloat[iblockIdx.x177{256}, iUS178{1}, iUR238{4}, ithreadIdx.x239{128}, iV237{2}, bS66{1}] ca_pos( 2 ) produce_pos( 2 )
 logical domain : (bS65{1}, bS66{1}, iS67{4096}, iS68{64})
 allocation domain : (bS65{1}, iS67{4096}, iS68{64}, bS66{1})
 contiguity: n t t n
  Split: iS67{4096} by factor 32 -> iS171{128}, iS172{32}
  Split: iS68{64} by factor 32 -> iS173{2}, iS174{32}
  Merge: iS172{32} and iS174{32} -> iS235{1024}
  Split: iS235{1024} by factor 2 -> iS236{512}, iV237{2}
  Merge: iS171{128} and iS173{2} -> iS175{256}
  Split: iS236{512} by factor 128 -> iUR238{4}, ithreadIdx.x239{128}
  Merge: bS65{1} and iS175{256} -> iS176{256}
  Split: iS176{256} by factor 1 -> iblockIdx.x177{256}, iUS178{1}
 loop domain : (iblockIdx.x177{256}, iUS178{1}, iUR238{4}, ithreadIdx.x239{128}, iV237{2}, bS66{1})
```

It is less than ideal that the problem with the merge resulted in
runtime failures and none of the lowing-time validations flagged
anything. However, this scheduling isn't invalid, except the
vectorization, although not ideal, so it isn't trivial to detect this at
lowering. For now, I added a simple sanity check.