Reorder maps for loop and logical domain in pointwise schedulers (#5237)
Pointwise scheduler uses `maybeReorderAsAllocationMap` to generate a map
for reordering loop and logical domain as the allocation domain. It
computes a permutation between loop and allocation domain. Loop and
logical domain are assumed to be the same, as is generally the case for
single-GPU programs at this stage.

For multi-GPU, so far this meant having any empty map. Recall, that the
`FinalizeMultideviceDomainsPass` reorders loop as allocation to avoid
existing limitations.

With #4381, loop and allocation can be different in order or
parallelization. This implies that the map generated to reorder loop as
allocation domain for multi-device fusions cannot be used for logical
domain. This PR generates different maps for logical and loop domain.

**Reordering logical as allocation domain:**

Allocation domain is assumed to post-dominate logical domain. We
canonicalize the allocation domain such that it is a permutation of
logical:
> When `allocation` is a split of `logical`, walks backwards from
`allocation` to `logical` to find a permutation of `logical` that
satisfies the order in `allocation`.

**Reordering loop as allocation domain:**

Loop domain is assumed to post-dominate allocation domain. We traverse
the transforms between allocation and loop domain such that it becomes a
permutation of loop domain. Caveat, this will not be the case if
allocation and loop are disjoint or allocation contains transforms
absent in loop domain. I am adding an assert to catch such cases early
and modify our implementation.